+++ bidon/src/main/java/org/bidon/sdk/config/DefaultAdapters.kt
@@ -21,6 +21,7 @@ enum class DefaultAdapters(val classPath: String) {
     MintegralAdapter(classPath = "org.bidon.mintegral.MintegralAdapter"),
     MobilefuseAdapter(classPath = "org.bidon.mobilefuse.MobileFuseAdapter"),
     MolocoAdapter(classPath = "org.bidon.moloco.MolocoAdapter"),
+    TaurusXAdapter(classPath = "org.bidon.taurusx.TaurusXAdapter"),
     UnityAdsAdapter(classPath = "org.bidon.unityads.UnityAdsAdapter"),
     VkAdsAdapter(classPath = "org.bidon.vkads.VkAdsAdapter"),
     VungleAdapter(classPath = "org.bidon.vungle.VungleAdapter"),