+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.12.0 (2025.**.**)
 ## Features:
 - [BDN-1011](https://appodeal.atlassian.net/browse/BDN-1011) Decoupled adapter version from sdk core version
+- [BDN-1052](https://appodeal.atlassian.net/browse/BDN-1052) Added TaurusX SDK 3.12.0 as ad adapter
 
 # 0.11.0 (2025.09.10)
 ## Features: